Implementing Single Responsibility Principle and separate amount for same product

+++ src/ProductInfo.java
@@ -1,12 +1,18 @@
-//This class ProductInfo is to store the product details name of the product, tax index that it belongs and the unit price of the product.
-//This class is used to create object and used as a dataType.
+/**
+ * ProductInfo : is to store product details (1.name of the product, 2.tax index that it belongs, 3.unit price of product and 4.Units of product available)
+ * it has
+ *  -product name.
+ *  -tax index to determine the tax category it belongs to.
+ *  -unit price of the product.
+ *  -units of that product available in shop.
+ */
 
 class ProductInfo {
     String name;
     int taxIndex;
-    int unitPrice;
+    double unitPrice;
     int unitsAvailable;
-    ProductInfo(String name, int index, int price, int unitsAvailable) {
+    ProductInfo(String name, int index, double price, int unitsAvailable) {
         this.name = name;
         this.taxIndex = index;
         this.unitPrice = price;
